Refactor with code review in many files
