+++ src/components/Board/Board.tsx
@@ -1,8 +1,7 @@
-import { useState } from 'react';
 import { Square } from '../Square/Square';
-import './board.css';
+import './Board.css';
 
-type SquareValue = 'X' | 'O' | null;
+export type SquareValue = 'X' | 'O' | null;
 
 interface BoardProps {
   xIsNext: boolean;
@@ -10,7 +9,33 @@ interface BoardProps {
   onPlay: (nextSquares: SquareValue[]) => void;
 }
 
-const Board: React.FC<BoardProps> = ({ xIsNext, squares, onPlay }) => {
+export const Board: React.FC<BoardProps> = ({ xIsNext, squares, onPlay }) => {
+  const calculateWinner = (squares: Array<SquareValue>) => {
+    const lines: number[][] = [
+      [0, 1, 2],
+      [3, 4, 5],
+      [6, 7, 8],
+      [0, 3, 6],
+      [1, 4, 7],
+      [2, 5, 8],
+      [0, 4, 8],
+      [2, 4, 6],
+    ];
+
+    for (let i = 0; i < lines.length; i++) {
+      const [a, b, c] = lines[i];
+      if (
+        squares[a] &&
+        squares[a] === squares[b] &&
+        squares[a] === squares[c]
+      ) {
+        return squares[a];
+      }
+    }
+
+    return null;
+  };
+
   const handleClick = (i: number) => {
     if (squares[i] || calculateWinner(squares)) {
       return;
@@ -22,16 +47,15 @@ const Board: React.FC<BoardProps> = ({ xIsNext, squares, onPlay }) => {
   };
 
   const winner = calculateWinner(squares);
-  let status;
-  if (winner) {
-    status = 'Winner: ' + winner;
-  } else {
-    status = 'Next player: ' + (xIsNext ? 'X' : 'O');
-  }
+
+  const status = winner
+    ? 'Winner: ' + winner
+    : 'Next player: ' + (xIsNext ? 'X' : 'O');
 
   return (
-    <div className="board-row">
+    <div className="board-container">
       <div className="status">{status}</div>
+
       <div className="grid">
         {Array(9)
           .fill(null)
@@ -45,69 +69,4 @@ const Board: React.FC<BoardProps> = ({ xIsNext, squares, onPlay }) => {
       </div>
     </div>
   );
-};
-
-export const Game = () => {
-  const [xIsNext, setXIsNext] = useState(true);
-  const [history, setHistory] = useState([Array(9).fill(null)]);
-  const [currentMove, setCurrentMove] = useState(0);
-  const currentSquares = history[currentMove];
-
-  const handlePlay = (nextSquares: SquareValue[]) => {
-    const nextHistory = [...history.slice(0, currentMove + 1), nextSquares];
-    setHistory(nextHistory);
-    setCurrentMove(nextHistory.length - 1);
-    setXIsNext(!xIsNext);
-  };
-
-  const jumpTo = (nextMove: number) => {
-    setCurrentMove(nextMove);
-    setXIsNext(nextMove % 2 === 0);
-  };
-
-  const moves = history.map((_squares: SquareValue[], move: number) => {
-    let description;
-    if (move > 0) {
-      description = 'Go to move #' + move.toString();
-    } else {
-      description = 'Go to game start';
-    }
-    return (
-      <li className="time" key={move}>
-        <button onClick={() => jumpTo(move)}>{description}</button>
-      </li>
-    );
-  });
-
-  return (
-    <div className="game">
-      <div className="game-board">
-        <Board xIsNext={xIsNext} squares={currentSquares} onPlay={handlePlay} />
-      </div>
-      <div className="game-info">
-        <ol>{moves}</ol>
-      </div>
-    </div>
-  );
-};
-
-const calculateWinner = (squares: Array<SquareValue>) => {
-  const lines: number[][] = [
-    [0, 1, 2],
-    [3, 4, 5],
-    [6, 7, 8],
-    [0, 3, 6],
-    [1, 4, 7],
-    [2, 5, 8],
-    [0, 4, 8],
-    [2, 4, 6],
-  ];
-
-  for (let i = 0; i < lines.length; i++) {
-    const [a, b, c] = lines[i];
-    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
-      return squares[a];
-    }
-  }
-  return null;
 };
