--- src/components/Board/board.css
@@ -1,43 +1,0 @@
-
-.board-row{
-    display:flex;
-    justify-content: center;
-    flex-direction: column;
-}   
-
-.grid{
-
-    display: grid;
-    grid-template-columns: repeat(3,6em);
-    justify-content: center;
-    margin-top: 2em;
-}
-
-ol{
-    list-style-type: none;
-
-}
-
-.time{
-    display: flex;
-    justify-content: center;
-    
-}
-
-.time button{
-    font-family: 'Press Start 2P', cursive;
-    padding: .2em 0;
-}
-
-.status{
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: .4em 0;
-    margin: .5em  0;
-    background-color: rgb(226, 232, 248);
-    text-align: center;
-    font-size: 40px;
-    font-family: 'Press Start 2P', cursive;
-    line-height: 1.1em;
-}
