+++ src/components/Square/square.css
@@ -1,7 +1,6 @@
-.square{
-    font-size: 50px;
-    font-weight: bold;
-    height: 100px;
-    width: 100px;
-    
+.square {
+  font-size: 50px;
+  font-weight: bold;
+  height: 100px;
+  width: 100px;
 }
